requirements.txt
django
psycopg2
gunicorn
whitenoise
factory_boy
